Reflect a gateway back again
The `bridge` plugin wants its bridge to have an IP to use for a
gateway IP (in particular if you've told it to act as a gateway). The
way to do this seems to be for the IPAM plugin to supply a gateway IP,
and since this is really in control of the operator, it needs to be in
config.